update ubuntu_build_deps.sh
 fix issue #387

+++ core/ubuntu_build_deps.sh
@@ -3,4 +3,6 @@
 sudo apt-get install -y gfortran libmysqlclient-dev mysql-client libcurl4-openssl-dev libboost-system-dev \
 libboost-filesystem-dev libboost-serialization-dev libboost-regex-dev
 
-sudo ln -s /usr/lib/x86_64-linux-gnu/libmysqlclient.so /usr/lib/x86_64-linux-gnu/libmysqlclient_r.so
+if [ ! -f "/usr/lib/x86_64-linux-gnu/libmysqlclient_r.so" ]; then
+    sudo ln -s /usr/lib/x86_64-linux-gnu/libmysqlclient.so /usr/lib/x86_64-linux-gnu/libmysqlclient_r.so
+fi
